--- src/main/java/com/pitchain/oauth/application/RequestOauthInfoService.java
@@ -1,28 +1,0 @@
-package com.pitchain.oauth.application;
-
-import com.pitchain.common.constant.OauthProvider;
-import com.pitchain.oauth.application.client.OauthClient;
-import com.pitchain.oauth.application.client.memberinfo.OauthMemberInfo;
-import com.pitchain.oauth.presentation.param.OauthParams;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-@Component
-public class RequestOauthInfoService {
-    private final Map<OauthProvider, OauthClient> clients;
-
-    public RequestOauthInfoService(List<OauthClient> clients) {
-        this.clients = clients.stream().collect(
-                Collectors.toUnmodifiableMap(OauthClient::oauthProvider, Function.identity()));
-    }
-
-    public OauthMemberInfo request(OauthParams oauthParams) {
-        OauthClient client = clients.get(oauthParams.oauthProvider());
-        String accessToken = client.getOauthLoginToken(oauthParams);
-        return client.getMemberInfo(accessToken);
-    }
-}
