+++ src/main/java/com/pitchain/oauth/application/handler/GoogleOauthHandler.java
@@ -1,41 +1,40 @@
-package com.pitchain.oauth.application.client;
+package com.pitchain.oauth.application.handler;
 
 import com.pitchain.common.apiPayload.ErrorStatus;
 import com.pitchain.common.constant.OauthProvider;
 import com.pitchain.common.exception.GeneralException;
-import com.pitchain.oauth.application.client.memberinfo.KakaoMemberInfo;
-import com.pitchain.oauth.application.client.memberinfo.OauthMemberInfo;
+import com.pitchain.oauth.application.memberinfo.GoogleMemberInfo;
+import com.pitchain.oauth.application.memberinfo.OauthMemberInfo;
 import com.pitchain.oauth.presentation.param.OauthParams;
-import com.pitchain.oauth.presentation.token.KakaoToken;
+import com.pitchain.oauth.presentation.token.GoogleToken;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
-import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
 @Slf4j
 @Component
-public class KakaoClient implements OauthClient{
-    @Value("${oauth.kakao.token_url}")
+public class GoogleOauthHandler implements OauthHandler {
+    @Value("${oauth.google.token_url}")
     private String token_url;
-    @Value("${oauth.kakao.user_url}")
+    @Value("${oauth.google.user_url}")
     private String user_url;
-    @Value("${oauth.kakao.grant_type}")
+    @Value("${oauth.google.grant_type}")
     private String grant_type;
-    @Value("${oauth.kakao.client_id}")
+    @Value("${oauth.google.client_id}")
     private String client_id;
-    @Value("${oauth.kakao.client_secret}")
+    @Value("${oauth.google.client_secret}")
     private String client_secret;
-    @Value("${oauth.kakao.redirect_uri}")
+    @Value("${oauth.google.redirect_uri}")
     private String redirect_uri;
 
     @Override
     public OauthProvider oauthProvider() {
-        return OauthProvider.KAKAO;
+        return OauthProvider.GOOGLE;
     }
 
     @Override
@@ -43,54 +42,50 @@ public class KakaoClient implements OauthClient{
         String url = token_url;
         log.debug("Authorization code: " + oauthParams.getAuthorizationCode());
 
-        //rest object 생성
         RestTemplate rt = new RestTemplate();
-        //헤더 생성
+        // 헤더 생성
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
 
-        //바디 생성
+        // 바디 생성
         MultiValueMap<String, String> body = oauthParams.makeBody();
         body.add("grant_type", grant_type);
         body.add("client_id", client_id);
         body.add("client_secret", client_secret);
         body.add("redirect_uri", redirect_uri);
 
-        //헤더 + 바디
+        // 헤더 + 바디 결합
         HttpEntity<MultiValueMap<String, String>> tokenRequest = new HttpEntity<>(body, headers);
         log.debug("Current httpEntity state: " + tokenRequest);
 
-        //소셜토큰 수신
-        KakaoToken kakaoToken = rt.postForObject(url, tokenRequest, KakaoToken.class);
-        log.debug("accessToken: " + kakaoToken);
+        // 토큰 수신
+        GoogleToken googleToken = rt.postForObject(url, tokenRequest, GoogleToken.class);
+        log.debug("accessToken: " + googleToken);
 
-        if (kakaoToken == null) {
-            log.error("kakao token을 정상적으로 가져오지 못했습니다.");
+        if (googleToken == null) {
+            log.error("google token을 정상적으로 가져오지 못했습니다.");
             throw new GeneralException(ErrorStatus._BAD_REQUEST);
         }
 
-        return kakaoToken.getAccess_token();
+        return googleToken.getAccess_token();
     }
 
     @Override
     public OauthMemberInfo getMemberInfo(String accessToken) {
         String url = user_url;
+        log.debug("Received token: " + accessToken);
 
-        //rest object 생성
+        // 요청 객체 생성
         RestTemplate rt = new RestTemplate();
 
-        //헤더 생성
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-        headers.add("Authorization", "Bearer " + accessToken);
+        // 헤더 생성
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setContentType(MediaType.APPLICATION_JSON);
+        httpHeaders.add("Authorization", "Bearer " + accessToken);
 
-        //바디 생성
-        LinkedMultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("property_keys", "[\"kakao_account.email\", \"kakao_account.profile\"]");
+        // Google 사용자 정보 요청 시 바디는 필요 없음
+        HttpEntity<String> memberInfoRequest = new HttpEntity<>(httpHeaders);
 
-        //헤더 + 바디
-        HttpEntity<LinkedMultiValueMap<String, String>> memberInfoRequest = new HttpEntity<>(body, headers);
-
-        return rt.postForObject(url, memberInfoRequest, KakaoMemberInfo.class);
+        return rt.postForObject(url, memberInfoRequest, GoogleMemberInfo.class);
     }
 }
